+++ sdk/resourcemanager/azure-resourcemanager-cosmos/src/main/java/com/azure/resourcemanager/cosmos/fluent/PercentileTargetsClient.java
@@ -17,7 +17,7 @@ public interface PercentileTargetsClient {
      * Retrieves the metrics determined by the given filter for the given account target region. This url is only for
      * PBS and Replication Latency data.
      *
-     * @param resourceGroupName Name of an Azure resource group.
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param accountName Cosmos DB database account name.
      * @param targetRegion Target region to which data is written. Cosmos DB region, with spaces between words and each
      *     word capitalized.
@@ -37,7 +37,7 @@ public interface PercentileTargetsClient {
      * Retrieves the metrics determined by the given filter for the given account target region. This url is only for
      * PBS and Replication Latency data.
      *
-     * @param resourceGroupName Name of an Azure resource group.
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param accountName Cosmos DB database account name.
      * @param targetRegion Target region to which data is written. Cosmos DB region, with spaces between words and each
      *     word capitalized.
@@ -57,7 +57,7 @@ public interface PercentileTargetsClient {
      * Retrieves the metrics determined by the given filter for the given account target region. This url is only for
      * PBS and Replication Latency data.
      *
-     * @param resourceGroupName Name of an Azure resource group.
+     * @param resourceGroupName The name of the resource group. The name is case insensitive.
      * @param accountName Cosmos DB database account name.
      * @param targetRegion Target region to which data is written. Cosmos DB region, with spaces between words and each
      *     word capitalized.