+++ sdk/resourcemanager/azure-resourcemanager-cosmos/src/main/java/com/azure/resourcemanager/cosmos/models/ArmResourceProperties.java
@@ -8,11 +8,26 @@ import com.azure.core.annotation.Immutable;
 import com.azure.core.management.Resource;
 import com.azure.core.util.logging.ClientLogger;
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import java.util.Map;
 
 /** The core properties of ARM resources. */
 @Immutable
 public class ArmResourceProperties extends Resource {
     @JsonIgnore private final ClientLogger logger = new ClientLogger(ArmResourceProperties.class);
+
+    /** {@inheritDoc} */
+    @Override
+    public ArmResourceProperties withLocation(String location) {
+        super.withLocation(location);
+        return this;
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public ArmResourceProperties withTags(Map<String, String> tags) {
+        super.withTags(tags);
+        return this;
+    }
 
     /**
      * Validates the instance.